gonna skip init calls for now, lets just setup a master template
