+++ packages/cli/package.json
@@ -39,6 +39,7 @@
   "dependencies": {
     "@startupkit/analytics": "workspace:*",
     "@startupkit/cms": "workspace:*",
+    "@startupkit/utils": "workspace:*",
     "@types/commander": "^2.12.2",
     "@types/node": "^22.2.0",
     "commander": "12.1.0",
