+++ packages/cli/tsconfig.json
@@ -11,6 +11,6 @@
     "resolveJsonModule": true,
     "skipLibCheck": true
   },
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "../utils/src/index.ts"],
   "exclude": ["node_modules"]
 }
