+++ packages/cms/src/init.ts
@@ -1,4 +1,12 @@
-import type { Initializer, InitializerQuestion } from "@startupkit/utils";
+import fs from "fs";
+import path from "path";
+import {
+  InitializerActions,
+  writeFile,
+  type Initializer,
+  type InitializerOptions,
+  type InitializerQuestion,
+} from "@startupkit/utils";
 
 const questions: InitializerQuestion[] = [
   {
@@ -11,8 +19,31 @@ const questions: InitializerQuestion[] = [
 
 type Answer = true | false;
 
-const init = async (cms: Answer) => {
-  console.log("Installing CMS", cms);
+// init() does the following:
+// 1. adds pages/blog/index.tsx
+// 2. adds pages/blog/sample-page.mdx
+// 3. adds @startupkit/analytics to the project
+
+const init = async (answers: Answer, opts: InitializerOptions) => {
+  const { cwd } = opts;
+
+  await fs.mkdirSync(path.join(cwd, "src/pages/blog"), { recursive: true });
+
+  await writeFile(
+    path.join(cwd, "src/blog/index.tsx"),
+    fs.readFileSync(path.join(__dirname, "../templates/index.tsx"), "utf8")
+  );
+  await writeFile(
+    path.join(cwd, "src/blog/sample-page.mdx"),
+    fs.readFileSync(
+      path.join(__dirname, "../templates/sample-page.mdx"),
+      "utf8"
+    )
+  );
+
+  return {
+    packages: ["@startupkit/cms"],
+  } satisfies InitializerActions;
 };
 
 export default {
